Update Problem_Using_For_Loop
For each integer  in the interval  (given as input) : [a,b]

If 1 <= n <= 9 , then print the English representation of it in lowercase. That is "one" for 1, "two" for 2, and so on.
Else if n > 9 and it is an even number, then print "even".
Else if n > 9 and it is an odd number, then print "odd".