Create Querying_The_Document
You will convert a raw text document into its component paragraphs, sentences and words. To test your results, queries will ask you to return a specific paragraph, sentence or word as described below.

Alicia is studying the C programming language at the University of Dunkirk and she represents the words, sentences, paragraphs, and documents using pointers:

A word is described by char* .
A sentence is described by char** . The words in the sentence are separated by one space (" "). The last word does not end with a space(" ").
A paragraph is described by char*** . The sentences in the paragraph are separated by one period (".").
A document is described by char**** . The paragraphs in the document are separated by one newline("\n"). The last paragraph does not end with a newline.